Renaming jac_method to SciPyFD
