+++ examples/all_softwares.ini
@@ -116,10 +116,10 @@ software: dfo
 
 # jac_method: This sets the options type of Jacobian used. Note: currently
 #             only numerical Jacobians are implemented.
-#             'scipy_fd' - denotes the use of SciPy's finite difference
+#             'SciPyFD' - denotes the use of SciPy's finite difference
 #                          Jacobian approximations
-#default 'scipy_fd'
-#jac_method: scipy_fd
+#default 'SciPyFD'
+#jac_method: SciPyFD
 
 # num_method: Sets the numerical method within used after setting the
 #             jac_method.
