+++ fitbenchmarking/results_processing/tests/test_support_page.py
@@ -8,7 +8,7 @@ except ImportError:
 import unittest
 
 from fitbenchmarking.parsing.fitting_problem import FittingProblem
-from fitbenchmarking.jacobian.scipy_fd_2point_jacobian import ScipyTwoPoint
+from fitbenchmarking.jacobian.SciPyFD_2point_jacobian import ScipyTwoPoint
 from fitbenchmarking.results_processing import support_page
 from fitbenchmarking.utils.fitbm_result import FittingResult
 from fitbenchmarking.utils.options import Options
